requirements.txt
black==22.8.0
Flask-Cors==3.0.9
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask==1.1.1
gunicorn==20.0.4
pylint==2.4.4
python-dotenv==0.12.0
Werkzeug==0.16.0